game-flow: merge GF_RunTitle to trx
