+++ src/tr1/game/game_flow/sequencer_misc.c
@@ -11,17 +11,6 @@
 #include <libtrx/log.h>
 #include <libtrx/utils.h>
 
-GF_COMMAND GF_RunTitle(void)
-{
-    Savegame_UnbindSlot();
-    GameStringTable_Apply(nullptr);
-    const GF_LEVEL *const title_level = GF_GetTitleLevel();
-    if (!Level_Initialise(title_level, GFSC_NORMAL)) {
-        return (GF_COMMAND) { .action = GF_EXIT_GAME };
-    }
-    return GF_ShowInventory(INV_TITLE_MODE);
-}
-
 GF_COMMAND GF_DoLevelSequence(
     const GF_LEVEL *const start_level, const GF_SEQUENCE_CONTEXT seq_ctx)
 {
